Moved Telemetry to Standard
