+++ src/VstsSyncMigrator.Core/Execution/OMatics/WorkItemLinkOMatic.cs
@@ -1,4 +1,5 @@
 ﻿using Microsoft.TeamFoundation.WorkItemTracking.Client;
+using MigrationTools;
 using System;
 using System.Diagnostics;
 using System.Linq;
